requirements.txt
# Core ML/DL packages optimized for RTX 4090
torch==2.5.1+cu124
torchvision==0.20.1+cu124
torchaudio==2.5.1+cu124
torch-geometric==2.6.1
triton==3.1.0

# Transformers and related packages
transformers==4.46.2
accelerate==1.1.1
peft==0.13.2
bitsandbytes==0.44.1
tokenizers==0.20.3
sentencepiece==0.2.0
safetensors==0.4.5

# Training and optimization
trl==0.12.1
datasets==3.1.0
evaluate==0.4.3

# NVIDIA CUDA packages (CUDA 12.4 for RTX 4090)
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127

# Memory optimization for large models
xformers==0.0.28.post2
flash-attn==2.6.3

# Unsloth for efficient fine-tuning
unsloth==2024.12

# Data processing
numpy==2.1.3
pandas==2.2.3
pyarrow==18.0.0
datasets==3.1.0

# Utilities
tqdm==4.67.1
rich==13.9.4
tyro==0.8.14
click==8.1.7
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
urllib3==2.2.3

# Scientific computing
networkx==3.4.2
nltk==3.9.1
joblib==1.4.2
psutil==6.1.0

# Development tools
setuptools==75.6.0
wheel==0.45.1
pip==24.3.1

# API clients
openai==1.57.0
huggingface-hub==0.26.2

# System dependencies
filelock==3.16.1
fsspec==2024.10.0
packaging==24.2
protobuf==4.25.5
pydantic==2.10.3
pydantic_core==2.27.1
typing_extensions==4.12.2

# HTTP/async
aiohttp==3.11.7
aiohappyeyeballs==2.4.4
httpx==0.28.1
httpcore==1.0.7
anyio==4.7.0
sniffio==1.3.1

# Misc utilities
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
idna==3.10
six==1.17.0
python-dateutil==2.9.0.post0
pytz==2024.2
tzdata==2024.2
markdown-it-py==3.0.0
mdurl==0.1.2
MarkupSafe==2.1.5
Jinja2==3.1.4
Pygments==2.18.0
pyparsing==3.2.0
sympy==1.13.3
mpmath==1.3.0
xxhash==3.5.0
yarl==1.18.3
multidict==6.1.0
frozenlist==1.5.0
aiosignal==1.3.1
dill==0.3.9
multiprocess==0.70.17
annotated-types==0.7.0
docstring_parser==0.16
h11==0.14.0
hf_transfer==0.1.8
jiter==0.8.2
shtab==1.7.1
